requirements.txt
albumentations==1.3.1
cityscapesscripts==2.2.2
dadaptation==3.1
imagecorruptions==1.1.2
imgviz==1.7.2
instaboostfast==0.1.2
ipdb==0.13.13
lightning==2.0.1
lion_pytorch==0.1.2
matplotlib==3.7.2
memory_profiler==0.61.0
mmcls==0.25.0
mmcv==2.0.0
mmengine==0.8.1
mmyolo==0.6.0
nibabel==5.1.0
numpy==1.24.1
onnxruntime==1.15.1
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
packaging==23.1
Pillow==9.3.0
Pillow==10.0.0
prettytable==3.8.0
psutil==5.9.5
psutil==5.9.0
pycocotools==2.0.6
rich==13.5.2
sahi==0.11.14
scikit_learn==1.3.0
scipy==1.11.2
segment_anything==1.0
setproctitle==1.3.2
Shapely==2.0.1
six==1.16.0
skimage==0.0
terminaltables==3.1.10
timm==0.9.5
torch==2.0.1+cu117
torchmetrics==0.11.4
torchvision==0.15.2+cu117
tqdm==4.65.0
